1. Finished test script.
2. Add predictor.
3. Fix possible underflow in utils.data_utils.normalize.

+++ src/utils/data_utils.py
@@ -34,7 +34,7 @@ def to_array(tensor):
         raise NotImplementedError
 
 def normalize(tensor):
-    return tensor / 255.0
+    return tensor.type(torch.float32) / 255.0
 
 def denormalize(tensor):
     return tensor * 255.0
